Fix build break of Arduino code.
Arduino code MUST have the main .ino file in a subfolder with the same
name.  By moving the code into a folder called Arduino the code would no
longer open in the Arduino IDE and so would not compile.

The Arduino code is now in a sub folder IRCameraBulbController to allow
for this.